Solved CDash lack-of-conversion warnings
git-svn-id: https://svn.opendtect.org/od/branches/od_osg@31373 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/visSurvey/visrandomtrackdisplay.cc
@@ -699,7 +699,7 @@ void RandomTrackDisplay::updatePanelStripPath()
 	if ( trcbids[trcidx] == knots_[knotidx] )
 	{
 	    pathcrds += Coord( knots_[knotidx].inl, knots_[knotidx].crl );
-	    mapping += trcidx;
+	    mapping += (float) trcidx;
 	    knotidx++;
 	}
     }
@@ -724,7 +724,7 @@ void RandomTrackDisplay::setPanelStripZRange( const Interval<float>& rg )
 {
     const StepInterval<float> zrg( rg.start, rg.stop, appliedZRangeStep() );
     panelstrip_->setZRange( zrg );
-    const Interval<float> mapping( 0, zrg.nrSteps() );
+    const Interval<float> mapping( 0, (float) zrg.nrSteps() );
     panelstrip_->setZRange2TextureMapping( mapping );
 
     if ( getUpdateStageNr() )
